requirements.txt
inky
font-fredoka-one
flask
paho-mqtt